Add retries to all event source api calls
Modifying and deleting an event source is an async
operation so it's possible that an update operation is
still in progress when you run something like `chalice delete`.
To account for this, we retry both calls if we get a
ResourceInUseException exception that says the event source mapping
is still in use.