Add chalice invoke command
New Chalice invoke command for easy invocation of deployed lambda
functions. Invocation is done by the Chalice logical name and the stdin
is read and used as the payload. The results are printed to the screen
and can render stack traces in a natural format making debugging
individual lambda functions much easier.