WIP
Add some log statements and a configure function in the cli layer to
make the logger respect --debug. Not sure what exactly is going to be
the best thing to debug. If we assume that the Package objects are
correctly implemented then what I have here is fine since they are
representative of what the various steps are working with. However if
they are broken it may be more useful to simply ls the working directory
between each step to see what we acutally have... or both which would be
bery noisy but point out any inconsistencies.