Updated check ordering to handle symlinks to nowhere properly
The problem was that symlinks to nowehere are not entered
into the cache table, but the routine took a None entry during
scan to mean a new file prior to checking whether the file
actually existed all.

Reordered handling so that OSError will occur before last_mtime
is actually checked.