Verify we see a debug log from chalice code
The existing test_debug_flag_enables_logging test
was relying on a debug message being present from botocore.
If you package an empty project with nothing in requirements.txt
you'll get:

```
$ chalice --debug package out5
2018-12-05 15:27:26,586 botocore.hooks [DEBUG] Changing event name from creating-client-class.iot-data to creating-client-class.iot-data-plane
2018-12-05 15:27:26,590 botocore.hooks [DEBUG] Changing event name from before-call.apigateway to before-call.api-gateway
2018-12-05 15:27:26,591 botocore.hooks [DEBUG] Changing event name from request-created.machinelearning.Predict to request-created.machine-learning.Predict
...
```

which all come from botocore when you create a client.
The issue is that those messages in botocore are only present in
versions of botocore after the "service id" switchover.  Our
version range dependency of botocore includes versions of botocore
before we switched the internal to user service id.  As a result
this test can fail if you have old (but allowed) version of
botocore installed.

The fix is to explicitly verify a debug message from within
the chalice code base so it's something we can control.  I added
a debug message to the packager that's in every code path
to ensure we see debug logs.