Change param model test to use swagger instead of API calls
There's sometimes a long period of eventual consistency where we're
not able to call `get_resources()` and have it return the resource id
of the route with the named capture (`/path/{name}`).

This commit changes the test to query for the swagger document, and
then verify that the name capture is modeled correctly in the
swagger document.

Given we use swagger to create the rest API in the first place,
I think this is an acceptable tradeoff.