Retry kms errors in lambda deploys
The kms key used to encrypt env vars (even if it's the default
key we don't explicitly configure) grants permission
to our IAM role, so there's times when it takes a while for
that to propagate.  This commit adds an additional regex to
check when determining if we should retry.