Fix invoke command stack trace reporting
In python 3.7 Lambda Functions the format of the response under
stackTrace has changed. Lambda returns the serialized output of
extract_tb. In earlier Pythons before 3.7, this was a 4-tuple which is
serialized as a list of 4 element lists. Python 3.7 introduced the
StackSummary and FrameSummary objects. These are serialized as a list of
Strings already formatted like a stack trace. The error handling code in
invoke has been updated to handle either case correctly.