Refactor waiting mechanism for websocket dropped messages test
It's possible for a message to be sent over a websocket, and then for
the websocket redeployment message drop integration test to execute a
scan on the DynamoDB table before the websocket message has traveled
through API Gateway -> Lambda far enough to make the DynamoDB PutItem
call. This means consistent read won't wait for it, and the list of
numbers read from the DynamoDB table will be missing the last entry.

The messages can also be processed out of order, meaning that the last
message being echoed back does not mean that all prior messages have
already been handled.