Fix packager to download missing sdists before trying to build them
The packager has an initial download step to cast a wide net trying to
download "an" artifact for each package. This is to build a full
dependency closure of everything we want to install. Because we want to
get something for every package we place no restrictions on what type of
artifact we want. Pip naturally prefers system-specific wheels when
available. After this download step we check for for any missing
compatible wheels, and try to use pip to download wheels that are
linux compatible. This catches the case where we weren't running our
package command from a linux machine, and there were platform-specific
wheel files available.

There are several cases for packages missing a compatible wheel file at
this point:

1) There were no wheel files available for download at all. In this
   case our initial download call would have fetched an sdist.
2) There were no linux compatible wheel files available, and no wheel
   files compatible with your local system. In this case the initial
   download will have downloaded the sdist instead of a wheel. The
   secondary download to try and force a compatible wheel will have also
   failed. Leaving us with just the initial sdist.
3) There were no linux compatible wheel files available, and there was a
   compatible one for your local system. In this case the initial
   download will have fetched the wheel that is compatible with your
   local system. And the second call to try and force a linux compatible
   download will have failed. Leaving you with just the incompatible
   wheel locally.

The next step after this is to try and build the sdist into a wheel
ourselves. Previous to this commit we were assuming we had an sdist
locally to build because case 1 and 2 are much more common. In case 3
the package would fail to build since the list of sdists to build is
empty. Resulting in a missing dependency error.

This commit fixes case 3 by doing a pip sdist-only download on any
package missing both a compatible wheel file and an sdist. Now when we
move on to the build step the sdist will be present.