Run tests without cond vars install
This will catch the issue we ran into where watchdog
was always being imported (#1355).

Rather than run this for every python version and double
our test matrix, I just updated this to run on the latest
python version.  The types of problems this will catch aren't
tied to a single python version so we only need to run this
test once.