Make current_request thread safe in local mode
This builds on https://github.com/aws/chalice/pull/1375
by ensuring we swap out the current request on the instance
of the running app.

I've also added a functional test to check for thread safety.

Long term, we should move to the forkingmixin, but it's too
drastic of a change to swap out for a bug fix so it will need
more vetting before we can switch over.

Fixes #759.