Restructure log following abstractions
This refactors how the log/tail code is organized.
There's a few edge cases we need to consider that
make this slightly complicated:

* Account for the fact that we still need to paginate our
  initial set of results if there's a next token
* Account for new events added that match the start time
  we're filtering on.  We need check on a per-eventid basis
  to properly filter out duplicate results.

Given the non-trivial nature of the code, I moved that logic
out of awsclient and into the logs module.  There's now a separate
abstraction for generating log events that lets us focus
on these abstractions separately from the low level API calls
datetime parsing code.

This is still a best-effort attempt.  We're not going to be able
to get every log message in real time, especially if the underlying
lambda function has a high rate of log event generation.

As part of this change, I've also added the option to filter
by a start time.  This uses the same values as the CLI's
"--since" option.