Split up quickstart into separate tutorials
The original "quickstart" was basically our README
and for a while was the only documentation we had.

We've also added a lot of features since the original
README was created, and as a result that page wasn't
really a "quickstart" anymore.

So I've split out the quickstart to just show how to deploy
a basic REST API, and moved the remaining content out into
separate tutorials.

I'd also split up the websocket tutorial into two separate
tutorials, the echo server and the chat server.