Handle symbols with multiple (shadowed) namespaces
In the case where a function is defined twice, e.g:

```
def foo():
    pass

def foo():
    x = 'bar'
```

We were incorrectly looking up the corresponding symbol
table for the function.  This is because we were doing
a name lookup on the children of the parent scope and returning
as soon as we found a match.  In the example above that would
always return the symbol table for the first function, even
if we were analyzing the second `foo` function.  This would
result in errors because we aren't able to find symbols we expect
to be defined in the current scope.

To fix this, I modified the child table lookup to also match up line
numbers if provided, so that way we can match scope by verifying that
the name matches in addition to matching the lineno of the FunctionDef
AST node with the lineno of the associated symtable.  As far as I can
tell, this is the best way to do this (there's a get_id() but that seems
to be the address of an internal struct type representing an AST and not
of the corresponding AST node we get from the ast module).