Only prefix url_prefix to route decorator
A slightly better fix would be to not pass in an extra
(undocumented) kwarg, but instead to use the `unused` arg
to pass in additional data.  However, in the small change
someone is relying on this behavior, I'd prefer to keep the
args as is, but just ensure we only pass in this param to
the route decorator, which is the only decorator that can
support this value.