Revamp deployment pipeline from 'generate-pipeline' command
The deployment pipeline generated from the ``generate-pipeline``
command hasn't been touched in several years.  In looking over
the code there's been notable changes in the underlying services
we use for the deployment pipeline:

* There's a new 0.2 version of the CodeBuild buildspec that has
  notable improvements from the 0.1 version buildspec.
* The way that CodeBuild manages language images has changed.
  Rather than having separate codebuild images for each runtime,
  you can now specify the runtime you need in your codebuild spec
  file (leading to issues such as #1412).  As a result you couldn't
  generate a deployment pipeline that worked with Python 3.8.
* CloudFormation added support for ssm/ssm-secure/secretsmanager.
  While we can't use ssm-secure for the GitHub OAuth token,
  secretsmanager is supported so we can more securely store our
  token instead of requiring it as a template parameter.

The output of `generate-pipeline` is in a grey area in terms of what to
support for backwards compatibility.  If the expectation is you can use
the output `generate-pipeline` to successfully set up a deployment
pipeline, then these changes are backwards compatible.  However,
the template parameters are changed as well as the CodeBuild image used
which could negatively impact customers.  As a result, I've decided
to make this an opt-in change (for now).  There's a new
`--pipeline-version` parameter where specifying `v2` will get you
the new features listed above.  The default value is `v1` which
will generate the same template as previous versions.
The other benefit of making this opt-in is we can restrict which
Python versions we support.  The new CodeBuild images only support
Python 3.7 and 3.8 so we can only support these new versions without
requiring fallbacks for earlier python version (you can just use
`v1` if you want to use older python versions).

Note that there was a bugfix needed when configuring a GitHub
repository as a source.  This was fixed in the `v1` template
as well as the `v2` template.