Make Request obj more consistent with lambda events
Existing event objects are able to access the original input event
dictionary as well as the lambda context.  The HTTP Request object
is different in that sense which makes it harder to write generic
middleware that needs access to the event object and the lambda
context.  This change adds that consistency by exposing this as
attributes on a request.

For backwards compat reasons, we have to name them other things,
so while the naming isn't consistent, it's at least now possible
to write generic middleware (with a type check for HTTP requests).