Support returning native python types with binary types
When using `*/*` for binary types, we'll always return binary
back to the user.  However, we previously required that you
have to use a `Response` return type if you wanted to provide
a JSON response.  With this change, you can now return a native
python type, and we'll automatically convert it to a binary
response type (i.e encode the type as json, then base64 encode
it and set the isBase64Encoded flag).

The current behavior is to fail with a ValueError saying that we
expected a ``bytes`` type instead of a ``str`` type, so this change
is flipping an existing error case to a success case.

See https://github.com/aws/chalice/issues/1501#issuecomment-696286897

Fixes #1501.