Fix regression with blueprint invocation
This regressed when middleware was introduced in Chalice (#1514).
It made the assumption that modifications could deferred to the
wrapping and modification of handlers until registration time.
However, the handler string for blueprints references the module
directly, e.g. `chalicelib.my_function` which means that the
app registration is bypassed entirely.

This is problematic in general and needs a follow up PR, but
will require a significant change from how blueprints currently
work.  For now, this puts back the original behavior and
instead allows for middleware to be set directly on the event
handler so that we don't need to wrap the blueprint handlers.

Fixes #1535