Configure websocket endpoint from rest API id
We were previously using the domainName to construct the endpoint URL.
If the user configures a custom domain name for the websocket API, we'll
use the custom domain name as the endpoint URL.  However, we always need
to use the auto-generated value from API gateway when constructing the
endpoint URL for the websocket client.

We have to keep the existing `.configure()` method as is, it's part of
the public API.  To fix this, I added a new method,
`.configure_from_api_id()` that constructs the endpoint given an
apiId and stage.  The websocket handler then uses this new
method when configuring your websocket API.

Fixes #1503