Add support for HTTP middleware catching exceptions
Originally proposed in #1541.

This adds a new ``ChaliceUnhandledException`` error that
can be raised from a view function.  Chalice explicitly
allows this error to propagate so that middleware for
REST APIs can catch and process this exception as needed.
This also allows you to create an app/middleware with
consistent error handling (you can create your own errors
that subclass from ``ChaliceUnhandledException``).

As part of this change, there were a few inconsistencies I fixed:

* Update the error log message to use the path instead of the view
  function.  In the context of middleware, the error may not have
  originated from the view function so the path is a more consistent
  name to use.
* Remove error handling about unknown routes.  This was old code
  from the early versions of Chalice that was dead code.  API Gateway
  will never send a request for an unknown URL (it will short circuit
  and send back a 403 forbidden).  Additionally, if this were to
  actually happen the exception we raised would propagate out of
  Lambda resulting in a 502 error response which isn't what we
  would want anyways.

Closes #1541.