Ensure only a single reference to a managed layer
When packaging a SAM application, ensure that there's only
a single reference to the managed layer for each Lambda function.
The issue was that the `resource.layers` was using the same
reference to a list from the loaded config file so mutating
this list to add in the managed layer was updating the list of
Lambda layers for all Lambda functions.  Now we make a copy of
this list to ensure each Lambda function gets a unique list
of layers and mutations are local to each Lambda function.

Fixes #1563.