Clear import cache in tests
Due to Chalice using relative paths when importing an app,
in Python 3.8.10, we'll use a cached version of a previous test case's
sys.path location which will result in import errors, for example see:

https://github.com/jamesls/chalice/runs/2529906754

The GitHub action for 3.8 upgraded from 3.8.9 to 3.8.10 which triggered
this issue.  It looks like this issue has been resolved upstream
(https://bugs.python.org/issue44070) but we still need a fix in Chalice
until a new version of 3.8 is released.  Probably not a bad idea to
clear the cache anyways if we're removing stuff from sys.path.