Support pip v22.x
Rather than opting in to each new pip version, we'll assume
that this new entry point will work going forward.  We're already
controlling the pip versions we support in our setup.py, so right now
we have to bump our pip dependency range *and* update this conditional.
Now, in the case that the import string stays the same, we just have
to bump the dependency range of pip.

Fixes #1887