Switch stale issue closing to github actions
Probot no longer works, so we need to switch over to GitHub actions.
This uses the same actions as other repos in the aws org.  I've
disabled the ancient issues, as my immediate needs are to get
parity with the old behavior, which was to close issuse that have
no received a response.

I've also set dry-run mode so I can have this run for a bit and
verify the actions it would take make sense.