Update built-in deployer to only remove log retention policy
We have to choose between being consistent with cloudformation
vs. what we think is the ideal behavior.  The original version
would delete the entire log group is the `log_retention_in_days`
property was removed, as the way that CFN models this is a separate
log group resource, so removing the config option means the resource
is no longer generated.

For the built-in deployer, we can detect when this config option
has been removed and only remove the retention policy (which is a
separate API call) instead of deleting the entire log group.