Update packaging integration tests to use latest major version
On the latest version of Python supported by Lambda, we want to make
sure that we're able to package the latest major version of popular
third party packages.  We can't expect that older package versions will
be forward compatible with newer versions of Python.

I've also bumped the `legacy_version` to be what the previous latest
version was.