Bump min version supported by Pandas tests
The latest issue is that installing the older version will now pull in a
version of numpy that doesn't have pre-built wheels which results in the
test attempting to build numpy which fails.

I've also updated the error message to print a more verbose error to
make it quicker to troubleshoot.