Fix new lines on rendered changelog
As part of this change I switched the template over to markdown.  This
lets all the recent changelog links to issues/PRs get properly linked
(they weren't being rendered as links in restructured text).

This issue was introduced in the latest release which pulled in the
latest version of jmeslog which changed the default template being used.
We now use an explicit template to control against this.