Fix distutils warning when packaging/deploying apps
If distutils is imported before pip outside of a `setup.py`, you see a
deprecation warning about distutils.  This was caused by a depdency we use,
`inquirer`, using a dependency that used distutils.
Upgrading to the latest major version of `inquirer` fixes this issue, which
no longer depends on the affected package.  As a result, we no longer see
this warning message when running `chalice package` or `chalice deploy`.

This triggered several other dependencies that needed to be updated, which
resulted in a few cosmetic code tweaks to make pylint happy.